# backend/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: API — Iglesia Simiente Santa
  version: 0.4.0
  description: |
    API REST del sitio web de la Iglesia Simiente Santa.

    En F1 expone el endpoint de estado del sistema (`/healthz`), público y sin
    autenticación. En F2 (acceso y gestión de usuarios) añade la sesión del
    panel, la gestión de cuentas y roles, la inicialización única de la puesta
    en marcha y la auditoría de solo lectura. En F3 (portada e información
    general) añade la portada pública sin autenticación (`GET /api/v1/portada`,
    solo contenido publicado y ya localizado con fallback `en → es` resuelto en
    el servidor), la descarga pública de imágenes (`GET /api/v1/media/{fileName}`)
    y la gestión del módulo desde el panel (`/api/v1/admin/portada`, permiso
    `portada`), con publicación por elemento.

    ## Formato uniforme de respuestas (FR-012)

    - **Éxito (2xx)**: el DTO documentado de la operación (sobre de éxito).
    - **Error (4xx/5xx)**: siempre `ErrorEnvelope` (sobre de error), también para
      los fallos previsibles (p. ej. la base de datos no conectada).

    Quien consume la API implementa el manejo de respuestas **una sola vez**:
    2xx → tipo generado de la operación; resto → `ApiError` con `ErrorEnvelope`.

    ## Códigos que la API puede devolver

    | HTTP | `error.code` | Cuándo |
    |---|---|---|
    | 200 | — | Operación exitosa (`GET /healthz` con la base de datos conectada, y el resto de operaciones con 2xx) |
...

...

    ## Seguridad

...
servers:
...
security:
  - sessionCookie: []
paths:
...
  /api/v1/portada:
    get:
      tags: [portada]
      operationId: getPortada
      summary: Portada pública con la información general publicada
      description: |
        Devuelve la información general **publicada** (FR-001…FR-007) resuelta al
        idioma pedido (FR-008/FR-009): cada campo traducible usa la versión en
        inglés si existe y, si no, la versión en español (idioma base). **Nunca**
        incluye contenido en borrador (FR-013/SC-002) y **omite por completo**
        las secciones sin elementos publicados (Q10/SC-012): cuando una clave de
        sección está presente, sus arrays nunca son `null` (§8.1.2).

        Sin autenticación. `Cache-Control: no-store`. Sin paginación: es un
        documento de una sola página (las colecciones están acotadas en el
        servidor: ≤50 servicios y ≤20 canales de WhatsApp).
      security: []
      parameters:
        - name: lang
          in: query
          required: false
          description: Idioma de resolución de contenidos. Por defecto `es` (idioma base, Decisión 8).
          schema:
            type: string
            enum: [es, en]
            default: es
      responses:
        "200":
          description: Información general publicada, localizada a `lang`.
          headers:
            Cache-Control:
              description: "Siempre `no-store` (SC-003: los cambios guardados se ven en la primera carga posterior)."
              schema:
                type: string
                enum: [no-store]
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortadaPublica"
        "400":
          $ref: "#/components/responses/Invalid"
        "500":
          $ref: "#/components/responses/Internal"
        "503":
          $ref: "#/components/responses/DatabaseUnavailable"
...
  /api/v1/admin/portada:
    get:
      tags: [portada-admin]
      operationId: getPortadaAdmin
      summary: Estado completo de la portada para el panel (incluye borradores)
      description: |
        Devuelve **todo** el contenido del módulo —identidad, «quiénes somos»,
        contacto, horario, canales de WhatsApp y redes— en **ambos idiomas** y
        con su `publicationState` por elemento (FR-011, US3 esc. 7), incluidos los
        borradores (este es el único lado que los ve; exige el permiso `portada`).
        Los singletons (identidad, quiénes somos, contacto) son `null` hasta el
        primer guardado. Cada colección va en su sobre `{items}` (§8.1.2) acotado
        en el servidor (≤50 servicios, ≤20 canales; redes: 1 por red del catálogo).
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Estado completo del módulo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortadaAdmin"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/Internal"
        "503":
          $ref: "#/components/responses/DatabaseUnavailable"
...
  /api/v1/admin/portada/whatsapp:
    post:
      tags: [portada-admin]
      operationId: createWhatsappChannel
      summary: Crea un canal de WhatsApp
      description: |
        Un canal tiene nombre/propósito (es obligatorio) y **un único destino**
        (FR-005/Q4): `kind=direct` + número de teléfono (se normaliza a dígitos
        con `+` opcional y el visitante abre `https://wa.me/<dígitos>`) o
        `kind=group` + URL `https` de `chat.whatsapp.com`/`wa.me` (FR-015).
        Un canal exactamente igual a otro (mismo `kind`, destino y nombre
        normalizados) se rechaza como duplicado (edge case de la spec). Máximo 20
        canales. Auditoría: `home.whatsapp.create` (FR-017).
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhatsappChannelInput"
      responses:
        "201":
          description: Canal creado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappChannelAdmin"
        "400":
          $ref: "#/components/responses/Invalid"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/admin/portada/whatsapp/{id}:
    patch:
      tags: [portada-admin]
      operationId: updateWhatsappChannel
      summary: Edita un canal de WhatsApp (y/o lo publica o retira)
      description: |
        Igual que `updateScheduleItem` para un canal de WhatsApp (FR-005/FR-013/FR-014).
        Auditoría *(analyze M5)*: un cambio de datos registra `home.whatsapp.update`;
        un cambio solo de estado registra `home.publish`/`home.unpublish`; un cambio
        que modifica **datos y estado** deja **ambas** filas (FR-017).
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/PortadaItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhatsappChannelPatch"
      responses:
        "200":
          description: Canal guardado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappChannelAdmin"
        "400":
          $ref: "#/components/responses/Invalid"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/Internal"
    delete:
      tags: [portada-admin]
      operationId: deleteWhatsappChannel
      summary: Elimina un canal de WhatsApp
      description: "Borrado físico. Auditoría: `home.whatsapp.delete` (FR-017)."
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/PortadaItemId"
      responses:
        "204":
          description: Canal eliminado (sin cuerpo).
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/admin/portada/redes:
    post:
      tags: [portada-admin]
      operationId: createSocialLink
      summary: Crea el enlace de una red social
      description: |
        Catálogo fijo (FR-006/Q5): `facebook`, `instagram`, `youtube`, `tiktok` o
        `spotify`, con **un solo enlace por red**. Una red fuera del catálogo se
        rechaza (`400`, `details.network`) y un segundo enlace para la misma red
        responde `409`. La URL debe ser `https` del dominio oficial de esa red.
        Auditoría: `home.social.create` (FR-017).
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SocialLinkInput"
      responses:
        "201":
          description: Enlace creado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialLinkAdmin"
        "400":
          $ref: "#/components/responses/Invalid"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/admin/portada/redes/{id}:
    patch:
      tags: [portada-admin]
      operationId: updateSocialLink
      summary: Edita el enlace de una red social (y/o lo publica o retira)
      description: |
        Igual que `updateScheduleItem` para un enlace de red (FR-006/FR-013/FR-014).
        Auditoría *(analyze M5)*: un cambio de datos registra `home.social.update`;
        un cambio solo de estado registra `home.publish`/`home.unpublish`; un cambio
        que modifica **datos y estado** deja **ambas** filas (FR-017).
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/PortadaItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SocialLinkPatch"
      responses:
        "200":
          description: Enlace guardado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialLinkAdmin"
        "400":
          $ref: "#/components/responses/Invalid"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/Internal"
    delete:
      tags: [portada-admin]
      operationId: deleteSocialLink
      summary: Elimina el enlace de una red social
      description: "Borrado físico. Auditoría: `home.social.delete` (FR-017)."
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/PortadaItemId"
      responses:
        "204":
          description: Enlace eliminado (sin cuerpo).
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ss_session
      description: |
        Sesión en servidor **en Redis** (D-A7 confirmada el 2026-10-04): cookie
        `ss_session` `HttpOnly` con un token opaco cuyo **SHA-256** es la clave
        en Redis (el token en claro nunca se guarda). Caduca por inactividad
        (30 min) o por vida absoluta (1 h desde el inicio de sesión); su
        `Max-Age` es la vida absoluta. Se emite en `POST /api/v1/auth/login` y se
        borra en logout y en toda revocación. Las peticiones no seguras con
        sesión exigen además la cabecera `X-CSRF-Token` con el valor de la
        cookie `csrf_token`.
  parameters:
...
    PortadaItemId:
      name: id
      in: path
      required: true
      description: Identificador (UUID) del elemento de la portada.
      schema:
        type: string
        format: uuid
  schemas:
...
    PortadaPublica:
      type: object
      required: [lang]
      description: |
        Información general publicada, resuelta a `lang` (fallback `en → es` por
        campo, FR-009). Las claves de sección están presentes **solo** si tienen
        elementos publicados (Q10/SC-012); si están presentes, sus arrays nunca
        son `null`. `identity`, `about` y `contact` son objetos únicos (FR-002,
        FR-003, FR-007).
      properties:
        lang:
          type: string
          enum: [es, en]
          description: Idioma en el que se resolvieron los contenidos.
        identity:
          $ref: "#/components/schemas/IdentityPublic"
        about:
          $ref: "#/components/schemas/AboutPublic"
        schedule:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleItemPublic"
        whatsapp:
          type: array
          items:
            $ref: "#/components/schemas/WhatsappChannelPublic"
        socials:
          type: array
          items:
            $ref: "#/components/schemas/SocialLinkPublic"
        contact:
          $ref: "#/components/schemas/ContactPublic"
      additionalProperties: false
...
    PortadaAdmin:
      type: object
      required: [identity, about, contact, schedule, whatsapp, socials]
      description: |
        Estado completo del módulo (FR-011), en **ambos idiomas** y con
        `publicationState` por elemento. Los singletons son `null` hasta el primer
        guardado. Cada colección va en su sobre `{items}` (§8.1.2).
      properties:
        identity:
          oneOf:
            - $ref: "#/components/schemas/IdentityAdmin"
            - type: "null"
        about:
          oneOf:
            - $ref: "#/components/schemas/AboutAdmin"
            - type: "null"
        contact:
          oneOf:
            - $ref: "#/components/schemas/ContactAdmin"
            - type: "null"
        schedule:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/ScheduleItemAdmin"
          additionalProperties: false
        whatsapp:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/WhatsappChannelAdmin"
          additionalProperties: false
        socials:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/SocialLinkAdmin"
          additionalProperties: false
      additionalProperties: false
...
    WhatsappChannelAdmin:
      type: object
      required: [id, nameEs, kind, destination, publicationState, sortOrder, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        nameEs:
          type: string
          minLength: 1
          maxLength: 120
        nameEn:
          type: [string, "null"]
          maxLength: 120
        kind:
          type: string
          enum: [direct, group]
        destination:
          type: string
          maxLength: 320
          description: |
            Destino único del canal (FR-005): teléfono normalizado (`direct`) o
            URL del grupo (`group`).
        publicationState:
          $ref: "#/components/schemas/PublicationState"
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
    WhatsappChannelInput:
      type: object
      required: [nameEs, kind, destination]
      description: |
        Alta de un canal (FR-005/FR-015). `kind=direct` exige teléfono con
        formato válido; `kind=group` exige URL `https` de `chat.whatsapp.com` o
        `wa.me`. El duplicado exacto (kind + destino + nombre normalizados) → 409.
        `nameEn` acepta `""`/espacios y se guarda como `null` (analyze I6).
      properties:
        nameEs:
          type: string
          minLength: 1
          maxLength: 120
        nameEn:
          type: [string, "null"]
          maxLength: 120
        kind:
          type: string
          enum: [direct, group]
        destination:
          type: string
          minLength: 1
          maxLength: 320
        publicationState:
          $ref: "#/components/schemas/PublicationState"
        sortOrder:
          type: integer
      additionalProperties: false
    WhatsappChannelPatch:
      type: object
      minProperties: 1
      description: Edición parcial de un canal (al menos un campo presente). `nameEn` acepta `""`/espacios y se guarda como `null` (analyze I6).
      properties:
        nameEs:
          type: string
          minLength: 1
          maxLength: 120
        nameEn:
          type: [string, "null"]
          maxLength: 120
        kind:
          type: string
          enum: [direct, group]
        destination:
          type: string
          minLength: 1
          maxLength: 320
        publicationState:
          $ref: "#/components/schemas/PublicationState"
        sortOrder:
          type: integer
      additionalProperties: false
...
    SocialLinkAdmin:
      type: object
      required: [id, network, url, publicationState, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        network:
          $ref: "#/components/schemas/SocialNetwork"
        url:
          type: string
          format: uri
          maxLength: 500
        publicationState:
          $ref: "#/components/schemas/PublicationState"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
    SocialLinkInput:
      type: object
      required: [network, url]
      description: |
        Alta del enlace de una red (FR-006/FR-015): red del catálogo y URL `https`
        del dominio oficial de esa red. Un segundo enlace para la misma red → 409.
      properties:
        network:
          $ref: "#/components/schemas/SocialNetwork"
        url:
          type: string
          format: uri
          minLength: 1
          maxLength: 500
        publicationState:
          $ref: "#/components/schemas/PublicationState"
      additionalProperties: false
    SocialLinkPatch:
      type: object
      minProperties: 1
      description: Edición parcial del enlace de una red (al menos un campo presente).
      properties:
        network:
          $ref: "#/components/schemas/SocialNetwork"
        url:
          type: string
          format: uri
          minLength: 1
          maxLength: 500
        publicationState:
          $ref: "#/components/schemas/PublicationState"
      additionalProperties: false
...
  responses:
    NotFound:
      description: Ruta no documentada. El fallback del router responde este sobre (la stdlib respondería texto).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: not_found
              message: Recurso no encontrado
...
    Internal:
      description: |
        Error inesperado. El mensaje es genérico a propósito (FR-013): el
        detalle interno queda registrado en el log estructurado con su
        `request_id` para diagnóstico.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: internal
              message: Error interno del servidor
    DatabaseUnavailable:
      description: |
        Servicio vivo, base de datos NO conectada (fallo previsible de
        `/healthz`; FR-004). La respuesta llega en ≤2 s por el timeout del ping.
      headers:
        Cache-Control:
          description: El estado nunca debe ser cacheado por intermediarios.
          schema:
            type: string
            enum: [no-store]
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: database_unavailable
              message: La base de datos no está conectada
              details:
                database: disconnected
    Invalid:
      description: |
        Datos inválidos o incompletos, o contraseña que no cumple la política
        (FR-010). `details` indica qué campo corregir; nunca expone información
        interna.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: invalid
              message: Revisa los datos del formulario
              details:
                email: no tiene un formato de correo válido
    Unauthenticated:
      description: |
        Sin sesión válida (expirada por inactividad, cerrada o inexistente) o
        credenciales de inicio de sesión incorrectas. El mensaje de login es
        genérico y **no** revela si la cuenta existe (FR-003).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: unauthenticated
              message: Correo o contraseña incorrectos
    Forbidden:
      description: |
        Autenticado pero sin permiso para la operación (FR-016), o cuenta
        desactivada al iniciar sesión (US1 esc. 3), o con la obligación de
        cambiar la contraseña antes de usar el panel (en cuyo caso
        `details.reason = "password_change_required"`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: forbidden
              message: No tienes permiso para realizar esta operación
    Conflict:
      description: |
        Conflicto con el estado actual: correo o nombre de rol duplicado (Q5),
        rol con cuentas asignadas al intentar eliminarlo (FR-017), inicialización
        ya ejecutada (FR-007) u operación que dejaría el panel sin administración
        (FR-008, en cuyo caso `details.reason = "admin_required"`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: conflict
              message: Ese correo ya está en uso
...